Rename contracts to new folders
