+++ src/Social/Contract/Interactions/InteractionsTrait.php
@@ -2,10 +2,9 @@
 
 declare(strict_types=1);
 
-namespace Kanvas\Packages\Social\Contract\Traits;
+namespace Kanvas\Packages\Social\Contract\Interactions;
 
 use Kanvas\Packages\Social\Contract\Users\UserInterface;
-use Kanvas\Packages\Social\Contract\Interactions\InteractionTypesInterface;
 
 trait InteractionsTrait
 {
